Fix gap after footer
The logo in the footer has a lot of padding to create the circle
glow effect around it. This glow bleeds out of the footer and
creates a gap at the end of the page. The `box-shadow` being `#fff`
isn't even visible in the gap. Hiding the `overflow` on
`div#footer` solved this problem.

Also removed extra whitespaces.